feat: capture onRequestResult
